requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
ruamel.yaml>=0.17.21
pydantic>=1.9.0
locust>=2.2.3
pymongo>=3.10.1
colorlog>=6.7.0